Added :NONAME and updated Forth programs and test code for version 1.7.1.
 On branch 1.x

 Changes to be committed:
	modified:   forth-src/forth2html.4th
	modified:   forth-src/fsl/complex-test.4th
	modified:   forth-src/fsl/extras/machar.4th
	modified:   forth-src/fsl/extras/zzeta.4th
	modified:   forth-src/games/chessboard.4th
	modified:   forth-src/games/pentomino.4th
	modified:   forth-src/games/react.4th
	modified:   forth-src/mini-oof-demo.4th
	new file:   forth-src/qm/epr-sim.4th
	modified:   forth-src/rawfloat-test.4th
	new file:   forth-src/struct-200x-ext.4th
	modified:   forth-src/system-test/coreplus.4th
	modified:   forth-src/system-test/fatan2-test.4th
	modified:   forth-src/system-test/fpzero-test.4th
	modified:   forth-src/system-test/ieee-arith-test.4th
	modified:   forth-src/system-test/ieee-fprox-test.4th
	modified:   forth-src/system-test/to-float-test.4th
	new file:   forth-src/textbox.4th
	modified:   forth-src/x11/cs-strings-x11.4th
	modified:   src/ForthVM.cpp
	modified:   src/ForthVM.h
	modified:   src/ForthWords.h
	modified:   src/Makefile
	modified:   src/vm32-common.s
	modified:   src/vmc.c

+++ src/ForthVM.h
@@ -1,6 +1,6 @@
 // ForthVM.h
 //
-// Copyright (c) 1996--2020, Krishna Myneni
+// Copyright (c) 1996--2021, Krishna Myneni
 //   <krishna.myneni@ccreweb.org>
 //
 // This software is provided under the terms of the GNU
@@ -51,7 +51,7 @@ void OpsPushDouble (double);
 int OpsCompileByte ();
 int OpsCompileInt ();
 int OpsCompileDouble ();
-void PrintVM_Error (int);
+void PrintVM_Error (long int);
 int ForthVM (vector<byte>*, long int**, byte**);
 
 // The following C++ functions have C linkage
